Bug fixing, editing documentation:
Fixed bug with `outputs` prop in `MfeOutletDirective`;

Fixed bug with `loaderDelay` in `MfeOutletDirective`, before the fix, the `loaderDelay` timer was first waited, and then the MFE was loaded.
The current behavior of `loaderDelay` is that the delay timer and MFE bundle loading start at the same time.
The `loaderDelay` property sets the minimum loader display time in ms;

Edited and improved documentation in README.md file;

`IMfeModuleRootOptions` interface renamed to `NgxMfeOptions`;

Property `delay` in the `NgxMfeOptions` renamed to `loaderDelay`;

`OPTIONS` injection token renamed to `NGX_MFE_OPTIONS`;